Add files via upload
The following files are for the web interface. the missing code is for reading raspberry pi IO lines to trigger a JavaScript function.
Dashboard.html is where the information and alerts will be displayed
index.html is where the login is. default username = "RGU HACK" and password = "rgu-hack01". signup button does nothing. otherwise this page produces no output.

AboutSite.html is information about the project and contributors.